requirements.txt
flask
faiss-cpu
numpy
pypdf
python-dotenv
google-generativeai